[ABW-1584] `offDeviceMnemonic` factor source kind (#532)

+++ Sources/Clients/FactorSourcesClientLive/FactorSourcesClient+Live.swift
@@ -15,7 +15,7 @@ extension FactorSourcesClient: DependencyKey {
 			await getProfileStore().profile.factorSources
 		}
 
-		let addOffDeviceFactorSource: AddOffDeviceFactorSource = { source in
+		let saveFactorSource: SaveFactorSource = { source in
 			try await getProfileStore().updating { profile in
 				guard !profile.factorSources.contains(where: { $0.id == source.id }) else {
 					throw FactorSourceAlreadyPresent()
@@ -30,17 +30,27 @@ extension FactorSourcesClient: DependencyKey {
 			factorSourcesAsyncSequence: {
 				await getProfileStore().factorSourcesValues()
 			},
-			addPrivateHDFactorSource: { privateFactorSource in
+			addPrivateHDFactorSource: { request in
+				let privateFactorSource = request.privateFactorSource
+				if privateFactorSource.kind == .device {
+					try await secureStorageClient.saveMnemonicForFactorSource(privateFactorSource)
+				}
+				let factorSourceID = privateFactorSource.id
 
-				try await secureStorageClient.saveMnemonicForFactorSource(privateFactorSource)
-				let factorSourceID = privateFactorSource.hdOnDeviceFactorSource.factorSource.id
-				do {
-					try await addOffDeviceFactorSource(privateFactorSource.hdOnDeviceFactorSource.factorSource)
-				} catch {
-					// We were unlucky, failed to update Profile, thus best to undo the saving of
-					// the mnemonic in keychain (if we can).
-					try? await secureStorageClient.deleteMnemonicByFactorSourceID(factorSourceID)
-					throw error
+				/// We only need to save olympia mnemonics into Profile, the Babylon ones
+				/// already exist in profile, and this function is used only to save the
+				/// imported mnemonic into keychain (done above).
+				if request.saveIntoProfile {
+					do {
+						try await saveFactorSource(privateFactorSource.factorSource)
+					} catch {
+						if privateFactorSource.kind == .device {
+							// We were unlucky, failed to update Profile, thus best to undo the saving of
+							// the mnemonic in keychain (if we can).
+							try? await secureStorageClient.deleteMnemonicByFactorSourceID(factorSourceID)
+						}
+						throw error
+					}
 				}
 
 				return factorSourceID
@@ -75,7 +85,7 @@ extension FactorSourcesClient: DependencyKey {
 					return nil // failed? to find any Olympia `.device` factor sources
 				}
 			},
-			addOffDeviceFactorSource: addOffDeviceFactorSource,
+			saveFactorSource: saveFactorSource,
 			getSigningFactors: { request in
 				assert(request.signers.allSatisfy { $0.networkID == request.networkID })
 				return try await signingFactors(
@@ -174,11 +184,12 @@ extension FactorSourceKind: Comparable {
 	fileprivate var signingOrder: Int {
 		switch self {
 		case .ledgerHQHardwareWallet: return 0
-		case .device:
-			// we want to sign with device last, since it would allow for us to stop using
-			// ephemeral notary and allow us to implement a AutoPurgingMnemonicCache which
-			// deletes items after 1 sec, thus `device` must come last.
-			return 1
+		case .offDeviceMnemonic: return 1
+
+		// we want to sign with device last, since it would allow for us to stop using
+		// ephemeral notary and allow us to implement a AutoPurgingMnemonicCache which
+		// deletes items after 1 sec, thus `device` must come last.
+		case .device: return 2
 		}
 	}
 }
